Move validation error interface
