+++ frontend/src/util/http-service.ts
@@ -1,6 +1,7 @@
 import axios, { AxiosRequestConfig, AxiosResponse } from "axios";
 import Cookies from "js-cookie";
 import _ from "lodash";
+import { ValidationErrorData } from "./data";
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 function camelizeKeys(obj: any): any {
@@ -65,29 +66,29 @@ export default class HttpService {
     validateStatus: statusCode => (statusCode >= 200 && statusCode < 300) || (statusCode >= 400 && statusCode < 500),
   });
 
-  static async get<TResponse, TResult = void>(url: string, successFn: (response: TResponse) => TResult, failureFn?: (response: ValidationErrorData) => TResult): Promise<TResult> {
+  static async get<TResponse, TResult = void>(url: string, successFn: (response: TResponse) => TResult, failureFn?: (response: ErrorResponse) => TResult): Promise<TResult> {
     return await this.performRequestFn(this._axios.get, url, successFn, failureFn);
   }
 
-  static async post<TResponse, TRequest, TResult = void>(url: string, data: TRequest, successFn: (response: TResponse) => TResult, failureFn?: (response: ValidationErrorData) => TResult): Promise<TResult> {
+  static async post<TResponse, TRequest, TResult = void>(url: string, data: TRequest, successFn: (response: TResponse) => TResult, failureFn?: (response: ErrorResponse) => TResult): Promise<TResult> {
     return await this.performDataRequestFn(this._axios.post, url, data, successFn, failureFn);
   }
 
-  static async put<TResponse, TRequest, TResult = void>(url: string, data: TRequest, successFn: (response: TResponse) => TResult, failureFn?: (response: ValidationErrorData) => TResult): Promise<TResult> {
+  static async put<TResponse, TRequest, TResult = void>(url: string, data: TRequest, successFn: (response: TResponse) => TResult, failureFn?: (response: ErrorResponse) => TResult): Promise<TResult> {
     return await this.performDataRequestFn(this._axios.put, url, data, successFn, failureFn);
   }
 
-  static async delete<TResponse, TRequest, TResult = void>(url: string, data: TRequest, successFn: (response: TResponse) => TResult, failureFn?: (response: ValidationErrorData) => TResult): Promise<TResult> {
+  static async delete<TResponse, TRequest, TResult = void>(url: string, data: TRequest, successFn: (response: TResponse) => TResult, failureFn?: (response: ErrorResponse) => TResult): Promise<TResult> {
     return await this.performDataRequestFn(this._axios.delete, url, data, successFn, failureFn);
   }
 
 
-  private static async performRequestFn<TResponse, TResult>(axiosRequestFn: AxiosRequestFn<TResponse | ValidationErrorData>, url: string, successFn: (response: TResponse) => TResult, failureFn?: (response: ValidationErrorData) => TResult): Promise<TResult> {
+  private static async performRequestFn<TResponse, TResult>(axiosRequestFn: AxiosRequestFn<TResponse | ErrorResponse>, url: string, successFn: (response: TResponse) => TResult, failureFn?: (response: ErrorResponse) => TResult): Promise<TResult> {
     const response = await axiosRequestFn(url);
     return this.processResponse(response, successFn, failureFn);
   }
 
-  private static async performDataRequestFn<TResponse, TRequest, TResult>(axiosDataRequestFn: AxiosDataRequestFn<TResponse | ValidationErrorData, TRequest>, url: string, data: TRequest, successFn: (response: TResponse) => TResult, failureFn?: (response: ValidationErrorData) => TResult): Promise<TResult> {
+  private static async performDataRequestFn<TResponse, TRequest, TResult>(axiosDataRequestFn: AxiosDataRequestFn<TResponse | ErrorResponse, TRequest>, url: string, data: TRequest, successFn: (response: TResponse) => TResult, failureFn?: (response: ErrorResponse) => TResult): Promise<TResult> {
     const config: AxiosRequestConfig = {
       headers: {
         'X-CSRFToken': this.getCsrfToken(),
@@ -99,7 +100,7 @@ export default class HttpService {
   }
 
 
-  private static processResponse<TResponse, TResult>(response: AxiosResponse<TResponse | ValidationErrorData>, successFn: (response: TResponse) => TResult, failureFn?: (response: ValidationErrorData) => TResult): TResult {
+  private static processResponse<TResponse, TResult>(response: AxiosResponse<TResponse | ErrorResponse>, successFn: (response: TResponse) => TResult, failureFn?: (response: ErrorResponse) => TResult): TResult {
     if (this.isResponseSuccess(response)) {
       return successFn(response.data);
     } else if (this.isResponseValidationErrorData(response)) {
@@ -114,11 +115,11 @@ export default class HttpService {
   }
 
 
-  private static isResponseSuccess<TResponse>(response: AxiosResponse<TResponse | ValidationErrorData>): response is AxiosResponse<TResponse> {
+  private static isResponseSuccess<TResponse>(response: AxiosResponse<TResponse | ErrorResponse>): response is AxiosResponse<TResponse> {
     return response.status >= 200 && response.status < 300;
   }
 
-  private static isResponseValidationErrorData<TResponse>(response: AxiosResponse<TResponse | ValidationErrorData>): response is AxiosResponse<ValidationErrorData> {
+  private static isResponseValidationErrorData<TResponse>(response: AxiosResponse<TResponse | ErrorResponse>): response is AxiosResponse<ErrorResponse> {
     return response.status >= 400 && response.status < 600 && 'errors' in response.data;
   }
 
@@ -134,9 +135,9 @@ type AxiosRequestFn<TResponse> = (url: string, config?: AxiosRequestConfig) => P
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 type AxiosDataRequestFn<TResponse, TRequest> = (url: string, data?: TRequest, config?: AxiosRequestConfig) => Promise<AxiosResponse<TResponse>>;
 
-type ValidationErrorData = {
+type ErrorResponse = {
   errors: {
     global?: string[],
-    validation?: Record<string, string[] | Record<string, string[]>>,
+    validation?: ValidationErrorData,
   },
 };
