--- frontend/src/views/surveyform/data/validation-error-data.ts
@@ -1,2 +1,0 @@
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-export type ValidationErrorData = Record<string, Record<string, any>>;
